requirements.txt
notion-client 
requests 
python-dotenv 
click